requirements.txt
selenium
pytest
pytest-html
pytest-xdist
webdriver-manager
allure-pytest